fix(macos-build): Build the god-team merge on macOS
The merge of god-team/main assumes MSVC and Win32 throughout, so neither
game target compiled or linked. Nothing here changes CRC-relevant code.

Win32 APIs replaced with portable equivalents:
- QueryPerformanceCounter in FrameRateLimit::reset
- CreateThread in W3D_TakeCompressedScreenshot
- InterlockedExchangePointer and InterlockedCompareExchangePointer,
  now provided by the interlocked adapter
- The machine fingerprint helpers in NGMP_Helpers: gethostuuid for the
  machine GUID, getifaddrs/AF_LINK for the MAC address, statfs for the
  volume serial

Discord Rich Presence returns false on macOS: it needs LoadLibraryA and
a discord-rpc DLL the bundle does not ship. Tracked separately.

std::utc_clock is not in libc++, so the six token-lifetime timestamps in
OnlineServices_Auth fall back to std::system_clock.

Include paths follow the qualified-include refactor: always.h, WWCommon.h
and the relative GeneralsOnline includes are now spelled in full.

The vanilla target was missing what Zero Hour got: the four observer font
MSG_META values with their CASE_LABELs, GameClient::step, and RTS_ZEROHOUR
fences around the observer font handling in CommandXlat.

NGMP_OnlineServices_LobbyInterface and RoomsInterface gain out-of-line
destructors. ~NGMP_OnlineServicesManager is inline in a header that only
forward-declares them, and deleting an incomplete type leaves clang
emitting an external destructor call that MSVC silently drops.

Also casts through size_t before narrowing a pointer in GadgetListBox,
and follows the audio refactor in MacOSAudioManager.

+++ Core/GameEngine/Source/GameClient/GUI/Gadget/GadgetListBox.cpp
@@ -2893,7 +2893,7 @@ void ApplyListBoxRowAnimation(GameWindow *window, Int rowIndex, Int rowHeight, I
 	Int animKey;
 	if (it->second == LIST_ROW_ANIM_ID)
 	{
-		Int itemID = (Int)GadgetListBoxGetItemData(window, rowIndex);
+		Int itemID = (Int)(size_t)GadgetListBoxGetItemData(window, rowIndex);
 		if (itemID <= 0)
 			return;
